controle-medico

+++ src/main/java/com/conexa/controlemedico/controllers/ControleMedicoController.java
@@ -15,7 +15,7 @@ import java.util.Optional;
 
 @RestController
 @CrossOrigin(origins = "*", maxAge = 3600)
-@RequestMapping("/api/v1/signup")
+@RequestMapping("/api/v1")
 public class ControleMedicoController {
 
     final ControleMedicoService medicoService;
@@ -24,7 +24,7 @@ public class ControleMedicoController {
         this.medicoService = medicoService;
     }
 
-    @PostMapping
+    @PostMapping("/signup")
     public ResponseEntity<Object> saveControleMedico(@RequestBody @Valid ControleMedicoDto medicoDto) {
         if(medicoService.existsByEmail(medicoDto.getEmail())) {
             return ResponseEntity.status(HttpStatus.CONFLICT).body("Este email já está em uso!");
@@ -35,12 +35,15 @@ public class ControleMedicoController {
         if(!Objects.equals(medicoDto.getConfirmacaoSenha(), medicoDto.getSenha())) {
             return ResponseEntity.status(HttpStatus.PRECONDITION_FAILED).body("Confirmação de senha incorreta!");
         }
+
+        medicoDto.setSenha(medicoService.getPasswordEncoder().encode(medicoDto.getSenha()));
+        medicoDto.setConfirmacaoSenha(medicoService.getPasswordEncoder().encode(medicoDto.getConfirmacaoSenha()));
         var medicoModel = new ControleMedicoModel();
         BeanUtils.copyProperties(medicoDto, medicoModel);
         return ResponseEntity.status(HttpStatus.CREATED).body(medicoService.save(medicoModel));
     }
 
-    @GetMapping
+    @GetMapping("/all")
     public ResponseEntity<List<ControleMedicoModel>> getAllControleMedico() {
         return ResponseEntity.status(HttpStatus.OK).body(medicoService.findAll());
     }
@@ -48,29 +51,69 @@ public class ControleMedicoController {
     @GetMapping("/{id}")
     public ResponseEntity<Object> getOneControleMedico(@PathVariable(value = "id") Long id){
         Optional<ControleMedicoModel> controleOptional = medicoService.findById(id);
+
         if (!controleOptional.isPresent()) {
             return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Registro não encontrado.");
         }
+
         return ResponseEntity.status(HttpStatus.OK).body(controleOptional.get());
     }
 
+    @GetMapping("/login")
+    public ResponseEntity<Object> medicoLogin(@RequestParam String email,
+                                              @RequestParam String senha) {
+        Optional<ControleMedicoModel> controleLogin = medicoService.findByEmail(email);
+
+        if (controleLogin.isEmpty()) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Cadastro inexistente.");
+        }
+
+        boolean controleSenha = medicoService.getPasswordEncoder().matches(senha, controleLogin.get().getSenha());
+        HttpStatus status = (controleSenha) ? HttpStatus.OK : HttpStatus.UNAUTHORIZED;
+        return ResponseEntity.status(status).body((controleSenha) ? "Login efetuado com sucesso!" : "Senha incorreta.");
+    }
+
     @DeleteMapping("/{id}")
-    public ResponseEntity<Object> deleteControleMedico(@PathVariable(value = "id") Long id) {
+    public ResponseEntity<Object> deleteControleMedico(@PathVariable(value = "id") Long id, @RequestParam String email,
+                                                       @RequestParam String senha) {
+        Optional<ControleMedicoModel> controleLogin = medicoService.findByEmail(email);
         Optional<ControleMedicoModel> controleOptional = medicoService.findById(id);
+        boolean controleSenha = medicoService.getPasswordEncoder().matches(senha, controleLogin.get().getSenha());
+
         if (!controleOptional.isPresent()) {
             return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Registro não encontrado.");
         }
+        if (controleSenha == false){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Email e senha não combinam.");
+        }
+        if(!Objects.equals(controleOptional.get(), controleLogin.get())) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Acesso negado!");
+        }
+
         medicoService.delete(controleOptional.get());
         return ResponseEntity.status(HttpStatus.OK).body("Registro apagado com sucesso!");
     }
 
     @PutMapping("/{id}")
     public ResponseEntity<Object> updateControleMedico(@PathVariable(value = "id") Long id,
-                                                       @RequestBody @Valid ControleMedicoDto medicoDto) {
+                                                       @RequestBody @Valid ControleMedicoDto medicoDto,
+                                                       @RequestParam String email,
+                                                       @RequestParam String senha
+                                                       ) {
+        Optional<ControleMedicoModel> controleLogin = medicoService.findByEmail(email);
         Optional<ControleMedicoModel> controleOptional = medicoService.findById(id);
+        boolean controleSenha = medicoService.getPasswordEncoder().matches(senha, controleLogin.get().getSenha());
+
         if (!controleOptional.isPresent()) {
             return ResponseEntity.status(HttpStatus.NOT_FOUND).body("Registro não encontrado.");
         }
+        if (controleSenha == false){
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Email e senha não combinam.");
+        }
+        if(!Objects.equals(controleOptional.get(), controleLogin.get())) {
+            return ResponseEntity.status(HttpStatus.UNAUTHORIZED).body("Acesso negado!");
+        }
+
         var medicoModel = new ControleMedicoModel();
         BeanUtils.copyProperties(medicoDto, medicoModel);
         medicoModel.setId(controleOptional.get().getId());
